refactor: separate out an ai module
