+++ image_text_retrieval/ai/clip.py
